requirements.txt
# The Continental Access Control System
# Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Password Hashing
bcrypt==4.1.1

# JWT Tokens
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Database
# SQLite is built into Python, no need to install

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Encryption
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0
